requirements.txt
pinecone-client
sentence-transformers
faiss-cpu  # or faiss-gpu if using GPU support
numpy
langchain
openai
